[WJZH-227] changed random integer member number value to long value.

+++ src/test/java/com/kurly/cloud/point/api/batch/PointOrderPublishBatchTest.java
@@ -146,7 +146,7 @@ public class PointOrderPublishBatchTest implements CommonTestGiven {
   }
 
   long randomOrderNumber() {
-    return new Random().nextInt();
+    return new Random().nextLong();
   }
 
   @Nested
